omap: add rpmsg backend device
Add an rpmsg device that wires virtio to the OMAP mailbox and
enabled A9<->M3 communications.

Use the mailbox payload to indicate which of the virtqueues is
triggered. In addition, a simpler STATE_CHANGE mode is also
supported.

Map the two vrings and the IPC buffers as noncacheable; use
arbitrarily-predefined physical addresses.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>