rpmsg: add remote-processor messaging virtio bus
Add a virtio IPC bus, which enables kernel users to communicate
with remote processors over shared memory using a simple messaging
protocol.

Assign a local address for every local endpoint that is created,
and bind it to the user's callback. Invoke that callback when a message
is received carrying the user's address as the destination addr.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>